requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langgraph>=0.0.10
openai>=1.10.0
faiss-cpu>=1.7.4
pydantic>=2.6.0
pdfplumber>=0.10.3
numpy>=1.26.0
python-multipart>=0.0.7
requests>=2.31.0
